device-posture/device-posture.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
'use strict';

function device_posture_test(func, name, properties) {
  promise_test(async t => {
    const originalPosture = navigator.devicePosture.type;
    t.add_cleanup(async () => {
      await test_driver.set_device_posture(originalPosture);
    });

    return func(t);
  }, name, properties);
}

// REFERENCE https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/http/tests/inspector-protocol/device-posture/device-posture-change-event.js
device_posture_test(async (t) => {
  const watcher = new EventWatcher(t, navigator.devicePosture, ['change']);
  const postures = ['folded', 'continuous', 'folded'];
  for (const posture of postures) {
    test_driver.set_device_posture(posture);
    await watcher.wait_for('change').then(() => {
      assert_equals(posture, navigator.devicePosture.type);
    });
  }
}, 'Device posture test: device-posture-change-event.js.');

// REFERENCE https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/http/tests/inspector-protocol/device-posture/device-posture-default.js
device_posture_test(async (t) => {
  assert_equals(navigator.devicePosture.type, 'continuous');
  await test_driver.set_device_posture('folded');
  assert_equals(navigator.devicePosture.type, 'folded')
}, 'Device posture test: device-posture-default.js.');

// REFERENCE https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/http/tests/inspector-protocol/device-posture/device-posture-event-listener.js
device_posture_test(async (t) => {
  assert_equals(navigator.devicePosture.type, 'continuous');

  const promise = new Promise(resolve => {
    navigator.devicePosture.addEventListener(
      'change',
      () => { resolve(navigator.devicePosture.type); },
      { once: true }
    );
  });
  await test_driver.set_device_posture('folded');
  assert_equals(await promise, 'folded')
}, 'Device posture test: device-posture-event-listener.js.');

// REFERENCE https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/http/tests/inspector-protocol/device-posture/device-posture-media-queries.js
device_posture_test(async (t) => {
  assert_equals(navigator.devicePosture.type, 'continuous');
  assert_true(matchMedia('(device-posture: continuous)').matches);

  const foldedMQL = window.matchMedia('(device-posture: folded)');
  const promise = new Promise(resolve => {
    foldedMQL.addEventListener(
      'change',
      () => { resolve(foldedMQL.matches); },
      { once: true }
    );
  });
  await test_driver.set_device_posture('folded');
  assert_true(await promise)
}, 'Device posture test: device-posture-media-queries.js.');
</script>
